Added mapped sequences to index.
Resolver returns random results for circular references when it enters the tree randomly, this is because it starts the journey at a different point each time and will join loop at random points which creates more or less results. To solve this, we parse in sequence, so the journey is always the same and always top down, the results will always be consistent.

Signed-off-by: Dave Shanley <dave@quobix.com>