Adding unit tests for console
The console won't change signature, flags may be added and changed but the basic structure will stay. This means console testing should be simple enough to make sure we run all the code. We're going to need an error code return flag for errors on rulesets to really validate the runs.

Signed-off-by: Dave Shanley <dave@quobix.com>