Source rendering is now in, looks terrible but it’s there.
The bloat is extreme on this report now, so we will have to truncate results over a few hundred violations. There is no value in seeing thousands of violations for a single rule. The pattern will be apparent; We can use this as a weighting mechanism (what should I tackle first?)

Signed-off-by: Dave Shanley <dshanley@splunk.com>