Added in new VacuumReport feature
Allows a complete report to be snapped in time. Can be compressed and minified as well, which makes it much more friendly for storage :) the origial spec is preserved within the report, so there is no loss of fidelity anywhere. There is no reason to re-run through the linter however, so it should load lightning fast, for giant specs we should be able to get to single digit ms.