Adding report rebuilding mechanism
This new mechanism will check a file to see if it's a report file. If so, it will decompress it and decode it as fast as possible. The report is then ready to feed into any renderer within vacuum. Replays over time! this is a foundational feature for a much larger vision.

pb33f.io

Signed-off-by: Dave Shanley <dave@quobix.com>