Added support for vacuum report replay to CUI
The console UI can now read in vacuum reports and replay them. It's slightly slower than the html-report, mainly because we need to re-index (poor design in the CUI). The CUI has got to be refactored, the sketch code is causing a lot of grief in my head now.

streamlined the HTML report replay code. Now it's ultra fast. 55ms to parse and load compressed k8s report

Signed-off-by: Dave Shanley <dave@quobix.com>