Modifed the sample plugins names
The names of the functions and the names of the structs are not connected, they don't need to align at all - however this is a little confusing when trying to understand the relationship and how a plugin works. mhanson's feedback here helped to fine tune the learning experience, Even though the names of the files, the structs and the function names are not connected - it makes sence to keep them aligned when being used as a sample, to make things easy to understand.