Sped up example handling exponentially
This damn rule has been the slowest since I wrote it. I don't like that I combined everything in here, but it is what it is for now. I managed to speed things up by running every schema operation async. This creates god knows how many thens of thousands more threads, but it completly unblocks the rule at scale.

This rule is now super fast. I could get faster, but it's topping the charts for right now. Timeout in schema resolving increased to 500ms as it was tripping at lowerspeeds and creating random result counts.