Refactored vacuum to use pb33f shared libraries
pb33f is now the source of truth for low level, shared models and concepts. The Resolver, Index and generic utilities have all been moved out. This means a smaller codebase for vacuum, and less moving parts for the application to manage. It will now become a sister of princess beef.