perf(docker): trim down image size
Using the build image as the final one retains things in the image
that's not needed at runtime, for example source code and build cache
including dependencies.

Additionally, the `golang:1.18` image contains the entire Go SDK as well
as other software needed for building things, such as compilers etc,
that are not needed at vacuum image runtime either.

Use `debian:bullseye-slim` as the base image to fix both issues. Brings
down size of the final image from 1.16GB to 107MB.

We could conceivably use an even smaller base image for the final one
than that, but we end up having some native lib dependencies in the
binary; thus `scratch` as basis won't work. And as that's the case, it's
safer to use the same basis as the `golang:1.18` build image does (it's
currently based on Debian Bullseye) rather than jump to something
completely different (such as e.g. Alpine based one).