fix(docker): run instructions, set image working dir to `/work`
Just `docker dshanley/vacuum lint` isn't enough, at least the OpenAPI
spec needs to be mounted in the container for it to be available.

Also, running a container off the image needs the `run` argument, and
`--rm` is most likely what most people want.

We now mount the base directory rather than just the top level spec file
provides for finding referenced schemas in case they are in the same
subdir tree. `/work` is set as the default working dir for the image,
and is where the tree with the spec should be mounted at.

While at it, consolidate/clarify instructions related to Docker Hub vs
GHCR.