Applied fix for #142 and some other tweaks (#143)
* Applied fix for #142 and some other tweaks

There was a bug in the `unused_definitions` function that was not looking in all polymorphic collections for use. Added the missing two checks to address #142

Also, there was a bug I found when testing rules, I noticed that when running without a version, rules are called multipled times. This behavior is incorrect, there is no way to be reliable or accurate if multiple
versions of rule for different versions are operating. Added a check for a missing version and format. Without this data, the rule is skipped.

Updated some tests to include version details in sample spec snippet code.

* Tuning CI/CD config

Disabling auto-tag for main push, not everything needs a release. Also bumping version on go.