change: altered behavior of operation_description function #174
The oepration description fucntion throws a mega-ton of warnings at the best of times, which can seem overwhelming. Also it gets upset even if there is no summary provided. So now the function is a little more forgiving. If there is no description AND no summary, then the rule will fire.

Also fixed unused_component change that was failing a test.