(feat) consistent ordering and output of html-report #190
An interesting use-case popped up where vacuum is being used as a 'diff check' tool to determine if output has changed. Some of the data was being rendered randomly, due to it not being sorted consistently. This change has been made to ensure that reach run will be identical in order and HTML will always result in identical output.

Also to ensure this is possible (and will validate against a sha256 hash), there is a new `-d` flag available for the `html-report` command. This has been added to the docs at: https://quobix.com/vacuum/commands/html-report/

Signed-off-by: Dave Shanley <dave@quobix.com>