(fix): html-report ordering was still flaky, should now be robust #190
Examples being the culprit, in large specs with lots of examples, some ordering was failing because there was not enough detail in the sort (it was only checking a couple of properties). Now sorting checks all the way down the model and applies a hash to paths and messages to ensure consistent ordering.

Testing against stripe, k8s, petstore and user supplied specs that were causing issues.

Signed-off-by: Dave Shanley <dave@quobix.com>