Addressed issue #583
A double dot (deep search) will cause gaps in a spectral path property, this has been fixed.

This is still not a full complete fix, because the JSON path for results is calculated by the doctor, and this looks like a straight passthrough from the rule, which needs further investigation.